+++ HISTORY.md
@@ -54,6 +54,7 @@
 * new "text costumes" library for generating costumes from letters or words of text
 * took out "b block" costume from catalog
 * added microphone "resolution" concept governing "bins" (buffer / bin sizes)
+* added microphone "resolution" settings to GUI
 
 ### 2019-03-10
 * Objects, Blocks, Threads: added microphone note and pitch detection